pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-extraction-env"
version = "0.1.0"
description = "OpenEnv environment for structured data extraction from unstructured text"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest", "openai"]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]